refactor: typo fix
